v12 rye

+++ App/Http/Controllers/HapusController.php
@@ -9,6 +9,10 @@ class HapusController extends Controller
 {
     public function destroy()
     {
+        if (Auth::user() == '') {
+            return view('auth.login');
+            exit();
+        }
         $data = DB::table('users')->delete();
         return view('user1', compact('data'));
     }
